requirements.txt
# Core dependencies
pdfplumber>=0.10.0      # PDF text and table extraction
pypdf>=3.0.0            # PDF parsing fallback
httpx>=0.25.0           # HTTP client for Claude API

# API server
fastapi>=0.104.0        # Web framework
uvicorn>=0.24.0         # ASGI server
python-multipart>=0.0.6 # File uploads

# Database
sqlalchemy>=2.0.0       # ORM
psycopg2-binary>=2.9.0  # PostgreSQL adapter
alembic>=1.12.0         # Database migrations

# Authentication
PyJWT>=2.8.0            # JWT token handling

# Email
sendgrid>=6.10.0        # Email delivery

# Production
gunicorn>=21.0.0        # Production WSGI server

# Optional: for better PDF handling
# pytesseract>=0.3.10   # OCR for scanned documents
# pdf2image>=1.16.0     # Convert PDF pages to images

# Development
pytest>=7.4.0           # Testing
pytest-asyncio>=0.21.0  # Async test support
